# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Playbook API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "https://api.playbooklab.com.br",
      "description": "Producao"
    },
    {
      "url": "http://localhost:8787",
      "description": "Ambiente local de desenvolvimento"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
...
  "paths": {
    "/v1/messages": {
      "post": {
        "tags": [
          "Mensagens"
        ],
        "summary": "Enviar mensagem",
        "description": "Envia uma mensagem numa conversa existente. Aceita JSON ou `multipart/form-data` com arquivos no campo `attachments` (ate 5, 15 MB somados); com anexo, `text` e opcional.\n\nConsome o limite diario `messages`.",
        "operationId": "sendMessage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              },
              "example": {
                "chat_id": "abc123",
                "text": "Ola! Tudo bem?"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "chat_id"
                ],
                "properties": {
                  "chat_id": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mensagem enviada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSentEnvelope"
                },
                "example": {
                  "ok": true,
                  "data": {
                    "message_id": "msg_abc123"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/AccountPaused"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/AccountNotReady"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      }
    },
    "/v1/invitations": {
      "post": {
        "tags": [
          "Convites"
        ],
        "summary": "Enviar convite",
        "description": "Envia um convite de conexao para um destinatario. O `provider_id` e o identificador do destinatario no LinkedIn.\n\nConsome o limite diario `invitations`.",
        "operationId": "sendInvitation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendInvitationRequest"
              },
              "example": {
                "provider_id": "ACoAAB1234567890",
                "message": "Oi! Gostaria de conectar."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Convite enviado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationSentEnvelope"
                },
                "example": {
                  "ok": true,
                  "data": {
                    "invitation_id": "inv_abc123"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/AccountPaused"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProfile"
          },
          "409": {
            "$ref": "#/components/responses/AccountNotReady"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      }
    },
    "/v1/chats": {
      "get": {
        "tags": [
          "Conversas"
        ],
        "summary": "Listar conversas",
        "description": "Lista suas conversas. Use o `id` de cada conversa como `chat_id` ao enviar mensagens. Suporta paginacao por cursor.",
        "operationId": "listChats",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Numero maximo de conversas a retornar.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Cursor de paginacao retornado por uma chamada anterior.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unread",
            "in": "query",
            "required": false,
            "description": "`true` so nao lidas; `false` so lidas.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "So itens anteriores a esta data (ISO 8601, ex.: `2026-09-01` ou `2026-09-01T12:00:00Z`).",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "So itens posteriores a esta data (ISO 8601).",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de conversas (ver ChatListEnvelope).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatListEnvelope"
                },
                "example": {
                  "ok": true,
                  "data": {
                    "items": [
                      {
                        "id": "chat_abc123",
                        "name": "Fulano de Tal",
                        "timestamp": "2026-08-01T12:34:56.000Z",
                        "unread_count": 2,
                        "archived": 0,
                        "attendee_provider_id": "ACoAAB1234567890"
                      }
                    ],
                    "cursor": null
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/AccountPaused"
          },
          "409": {
            "$ref": "#/components/responses/AccountNotReady"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      },
      "post": {
        "tags": [
          "Conversas"
        ],
        "summary": "Iniciar conversa",
        "description": "Abre uma conversa nova com uma pessoa (ou retoma a existente) e envia a primeira mensagem. Para quem nao e sua conexao, use `inmail: true` (gasta credito de InMail). Para falar com uma pagina de empresa, informe `topic`.\n\nConsome o limite diario `messages`.",
        "operationId": "startChat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "provider_id"
                ],
                "properties": {
                  "provider_id": {
                    "type": "string",
                    "description": "`provider_id` da pessoa ou id da empresa."
                  },
                  "text": {
                    "type": "string",
                    "maxLength": 8000,
                    "description": "Obrigatorio sem anexo."
                  },
                  "inmail": {
                    "type": "boolean",
                    "description": "Enviar como InMail."
                  },
                  "subject": {
                    "type": "string",
                    "maxLength": 200,
                    "description": "Assunto (InMail)."
                  },
                  "topic": {
                    "type": "string",
                    "enum": [
                      "service_request",
                      "request_demo",
                      "support",
                      "careers",
                      "other"
                    ],
                    "description": "Obrigatorio para falar com empresa."
                  },
                  "invitation_id": {
                    "type": "string",
                    "description": "Para responder a quem te convidou, antes de aceitar o convite."
                  }
                }
              },
              "example": {
                "provider_id": "ACoAAB1234567890",
                "text": "Oi! Vi seu post sobre prospeccao e queria trocar uma ideia."
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider_id": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  },
                  "inmail": {
                    "type": "string"
                  },
                  "subject": {
                    "type": "string"
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Ate 5 arquivos, 15 MB somados."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "data"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "chat_id": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "message_id": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      }
                    }
                  }
                },
                "example": {
                  "ok": true,
                  "data": {
                    "chat_id": "Ab12Cd34Ef56Gh78Ij90Kl",
                    "message_id": "Mn12Op34Qr56St78Uv90Wx"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/AccountPaused"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResource"
          },
          "409": {
            "$ref": "#/components/responses/AccountNotReady"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "Sua chave de API, no formato `lk_live_...`. Enviada em toda requisicao no header `X-API-KEY`."
      }
    },
    "schemas": {
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "chat_id",
          "text"
        ],
        "additionalProperties": false,
        "properties": {
          "chat_id": {
            "type": "string",
            "description": "Id da conversa existente (obtido em `GET /v1/chats`)."
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "description": "Conteudo da mensagem."
          }
        }
      },
      "SendInvitationRequest": {
        "type": "object",
        "required": [
          "provider_id"
        ],
        "additionalProperties": false,
        "properties": {
          "provider_id": {
            "type": "string",
            "description": "Identificador do destinatario no LinkedIn."
          },
          "message": {
            "type": "string",
            "description": "Mensagem opcional que acompanha o convite."
          }
        }
      },
      "MessageSentEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "required": [
              "message_id"
            ],
            "properties": {
              "message_id": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Identificador da mensagem enviada."
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Requisicao invalida: campo obrigatorio ausente, formato errado, JSON malformado, cursor invalido, ou filtro recusado pelo LinkedIn (`invalid_request`, com `reason`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "missing_field": {
                "value": {
                  "error": "missing_chat_id"
                }
              },
              "invalid_json": {
                "value": {
                  "error": "invalid_json"
                }
              },
              "invalid_cursor": {
                "value": {
                  "error": "invalid_cursor"
                }
              },
              "invalid_request": {
                "value": {
                  "error": "invalid_request",
                  "reason": "invalid_parameters"
                }
              },
              "unknown_filter": {
                "value": {
                  "error": "unknown_filter",
                  "filter": "hiring_project"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Chave ausente, invalida ou revogada.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "missing_api_key": {
                "value": {
                  "error": "missing_api_key"
                }
              },
              "invalid_api_key": {
                "value": {
                  "error": "invalid_api_key"
                }
              }
            }
          }
        }
      },
      "AccountPaused": {
        "description": "Sua conta esta pausada porque o pagamento da assinatura esta em atraso. Ela volta sozinha assim que o Pix for confirmado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "account_paused"
            }
          }
        }
      },
      "AccountNotReady": {
        "description": "Sua chave e valida, mas nao ha LinkedIn ativo na conta: ou a sessao caiu (`account_disconnected`, reconecte pelo painel) ou nenhum LinkedIn foi conectado ainda (`linkedin_not_connected`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "account_disconnected": {
                "value": {
                  "error": "account_disconnected"
                }
              },
              "linkedin_not_connected": {
                "value": {
                  "error": "linkedin_not_connected"
                }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Limite diario da acao atingido, ou teto de tentativas do dia (`reason: too_many_attempts`). Aguarde a virada da janela (header `Retry-After` traz os segundos restantes).",
        "headers": {
          "Retry-After": {
            "description": "Segundos ate a proxima janela.",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "rate_limited",
              "action": "messages",
              "limit": 80,
              "retry_after": 3600
            }
          }
        }
      },
...
      "NotFound": {
        "description": "O recurso informado nao existe para a sua conta (id inexistente ou de outra conta).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "not_found"
            }
          }
        }
      },
...
      "PayloadTooLarge": {
        "description": "Corpo grande demais: anexos acima de 15 MB somados (`attachments_too_large`) ou JSON acima de 1 MB (`body_too_large`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "attachments_too_large",
              "max_bytes": 15728640
            }
          }
        }
      }
    }
  },
...
}